refactor(taskflow): lift cross-feature widgets into ui
Move Avatar, MarkdownView, MarkdownEditor from ui.components to ui package;
update all import sites tree-wide (AccountRow, KanbanCard, CardDetailScreen,
ProfileScreen, App).
